LIBRARY: fixed compilation on mac
/Users/runner/work/cubiquity/cubiquity/src/library/geometry.h:827:24: error: non-constant-expression cannot be narrowed from type 'result_type' (aka 'double') to 'float' in initializer list [-Wc++11-narrowing]
                        Vector3f result = { randX(eng), randY(eng), randZ(eng) };

LIBRARY: fixed compilation on macos

LIBRARY: fixed compilation on macos

+++ src/application/commands/test/test_rendering.cpp
@@ -51,7 +51,7 @@ bool testRasterization()
 
 		for (int ct = 0; ct < 5000; ct++)
 		{
-			Vector2i centre({ simple_rand() % maskSize, simple_rand() % maskSize });
+			Vector2i centre{ (int)(simple_rand() % maskSize), (int)(simple_rand() % maskSize) };
 
 			// Note that the base polygon has size of approx two (-1.0 to + 1.0)
 			// FIXME - It would be nice to test with zero-size (or very timy) polygons, but
